Moved session code to its own module.
